requirements.txt
absl-py==1.4.0
accelerate==1.0.1
aiofiles==22.1.0
aiosqlite==0.18.0
annotated-types==0.7.0
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
attrs==22.2.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.11.2
black==24.8.0
bleach==6.0.0
brotlipy==0.7.0
cachetools==5.3.0
certifi==2021.5.30
cffi @ file:///tmp/build/80754af9/cffi_1625807838443/work
chardet @ file:///tmp/build/80754af9/chardet_1607706746162/work
charset-normalizer==3.1.0
click==8.1.7
cmake==3.26.0
coloredlogs==15.0.1
comm==0.1.2
conda==4.10.3
conda-package-handling @ file:///tmp/build/80754af9/conda-package-handling_1618262148928/work
contourpy==1.0.7
cryptography @ file:///tmp/build/80754af9/cryptography_1616769286105/work
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
elastic-transport==8.15.1
elasticsearch==8.16.0
executing==1.2.0
fastapi==0.115.6
fastjsonschema==2.16.3
filelock==3.10.0
flatbuffers==24.3.25
fonttools==4.39.0
fqdn==1.5.1
fsspec==2024.10.0
google-auth==2.16.2
google-auth-oauthlib==0.4.6
graphsurgeon @ file:///root/TensorRT-8.6.1.6/graphsurgeon/graphsurgeon-0.4.6-py2.py3-none-any.whl
grpcio==1.51.3
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.26.5
humanfriendly==10.0
idna @ file:///home/linux1/recipes/ci/idna_1610986105248/work
importlib-metadata==6.0.0
importlib-resources==5.12.0
ipykernel==6.21.3
ipython==8.11.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
jiter==0.8.2
joblib==1.4.2
json5==0.9.11
jsonpointer==2.3
jsonschema==4.17.3
jupyter-client==8.0.3
jupyter-core==5.2.0
jupyter-events==0.6.3
jupyter-server==2.4.0
jupyter-server-fileid==0.8.0
jupyter-server-terminals==0.4.4
jupyter-server-ydoc==0.6.1
jupyter-ydoc==0.2.3
jupyterlab==3.6.1
jupyterlab-language-pack-zh-cn==3.6.post0
jupyterlab-pygments==0.2.2
jupyterlab-server==2.20.0
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
lit==15.0.7
llmlingua==0.2.2
Mako==1.3.8
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
milvus-lite==2.4.10
mistune==2.0.5
mpmath==1.3.0
mypy-extensions==1.0.0
mysql-connector-python==9.0.0
nbclassic==0.5.3
nbclient==0.7.2
nbconvert==7.2.10
nbformat==5.7.3
nest-asyncio==1.5.6
networkx==3.0
nltk==3.9.1
notebook==6.5.3
notebook-shim==0.2.2
numpy==1.24.2
oauthlib==3.2.2
onnx==1.14.0
onnx-graphsurgeon @ file:///root/TensorRT-8.6.1.6/onnx_graphsurgeon/onnx_graphsurgeon-0.3.12-py2.py3-none-any.whl
onnxruntime==1.19.2
openai==1.57.4
packaging==23.0
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.12.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pkgutil-resolve-name==1.3.10
platformdirs==3.1.1
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==4.22.1
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pycuda==2024.1.2
pydantic==2.10.3
pydantic-core==2.27.1
Pygments==2.14.0
PyJWT==2.8.0
pymilvus==2.5.0
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1608057966937/work
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
pytools==2024.1.14
pytz==2022.7.1
PyYAML==6.0
pyzmq==25.0.1
regex==2024.11.6
requests==2.28.2
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
ruamel-yaml-conda @ file:///tmp/build/80754af9/ruamel_yaml_1616016699510/work
safetensors==0.4.5
seaborn==0.13.2
Send2Trash==1.8.0
six @ file:///tmp/build/80754af9/six_1623709665295/work
sniffio==1.3.0
soupsieve==2.4
stack-data==0.6.2
starlette==0.41.3
supervisor==4.2.5
sympy==1.11.1
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorrt @ file:///root/TensorRT-8.6.1.6/python/tensorrt-8.6.1-cp38-none-linux_x86_64.whl
terminado==0.17.1
tiktoken==0.7.0
tinycss2==1.2.1
tokenizers==0.20.3
tomli==2.0.1
torch @ http://download.pytorch.org/whl/cu118/torch-2.0.0%2Bcu118-cp38-cp38-linux_x86_64.whl
torchvision @ http://download.pytorch.org/whl/cu118/torchvision-0.15.1%2Bcu118-cp38-cp38-linux_x86_64.whl
tornado==6.2
tqdm @ file:///tmp/build/80754af9/tqdm_1625563689033/work
traitlets==5.9.0
transformers==4.46.3
triton==2.0.0
typing-extensions==4.12.2
tzdata==2024.2
uff @ file:///root/TensorRT-8.6.1.6/uff/uff-0.6.9-py2.py3-none-any.whl
ujson==5.10.0
uri-template==1.2.0
urllib3 @ file:///tmp/build/80754af9/urllib3_1625084269274/work
uvicorn==0.32.1
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.2.3
widgetsnbextension==4.0.5
y-py==0.5.9
ypy-websocket==0.8.2
zhipuai==2.1.5.20241204
zipp==3.15.0
